+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   [#36](https://github.com/Ch00k/orthanc-cli/pull/36)
 * Allow configuring modification with command-line options (in addition to the config file)
   [#37](https://github.com/Ch00k/orthanc-cli/pull/37)
+* Implemented search within Orthanc [#38](https://github.com/Ch00k/orthanc-cli/pull/38)
 
 ## [0.1.1](https://github.com/Ch00k/orthanc-cli/compare/0.1.0...0.1.1) 2020-12-20